pyproject.toml
[project]
name = "rec-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.11",
    "django-restframework>=0.0.1",
    "djangorestframework>=3.16.1",
    "faiss-cpu>=1.13.2",
    "matplotlib>=3.10.8",
    "mlflow>=3.10.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pymysql>=1.1.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.2",
    "shap>=0.49.1",
    "sqlalchemy-utils>=0.42.1",
    "sqlmodel>=0.0.37",
    "zenml>=0.93.3",
]
